replace uint8_t with uint16_t

+++ include/vcf.h
@@ -45,7 +45,7 @@ public:
 
     bool pos_in_range(const uint32_t, const uint32_t, const std::string &) const;
 
-    void genotype(const uint32_t &, const float &, const uint8_t, const uint32_t &min_allele_covg,
+    void genotype(const uint32_t &, const float &, const uint16_t, const uint32_t &min_allele_covg,
                       const float &min_fraction_allele_covg, const uint32_t &min_site_total_covg,
                       const uint32_t &min_site_diff_covg, bool snps_only);
